fix -h command return code and start to split up default.py
work on issue #79 , #80 and #81
